requirements.txt
Flask==3.1.2
gunicorn==23.0.0
pydantic==2.11.7
pytest==8.4.2
requests==2.32.5
google-genai>=1.61.0